# api/swagger/loms/v1/loms.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ozon Univer Golang (loms)",
    "version": "0.0.1"
  },
  "tags": [
    {
      "name": "LomsService"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/order/cancel": {
      "post": {
        "operationId": "LomsService_OrderCancel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1OrderCancelResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1OrderCancelRequest"
            }
          }
        ],
        "tags": [
          "LomsService"
        ]
      }
    },
    "/order/create": {
      "post": {
        "operationId": "LomsService_OrderCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1OrderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1OrderCreateRequest"
            }
          }
        ],
        "tags": [
          "LomsService"
        ]
      }
    },
    "/order/info": {
      "post": {
        "operationId": "LomsService_OrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1OrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1OrderInfoRequest"
            }
          }
        ],
        "tags": [
          "LomsService"
        ]
      }
    },
    "/order/pay": {
      "post": {
        "operationId": "LomsService_OrderPay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1OrderPayResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1OrderPayRequest"
            }
          }
        ],
        "tags": [
          "LomsService"
        ]
      }
    },
    "/stocks": {
      "post": {
        "operationId": "LomsService_GetStocks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetStocksResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetStocksRequest"
            }
          }
        ],
        "tags": [
          "LomsService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetStocksRequest": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1GetStocksResponse": {
      "type": "object",
      "properties": {
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1OrderCancelRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1OrderCancelResponse": {
      "type": "object"
    },
    "v1OrderCreateItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1OrderCreateRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1OrderCreateItem"
          }
        }
      }
    },
    "v1OrderCreateResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1OrderInfoItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1OrderInfoRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1OrderInfoResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1OrderInfoItem"
          }
        }
      }
    },
    "v1OrderPayRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1OrderPayResponse": {
      "type": "object"
    }
  }
}